Separate interceptor recorder and reader in files
